+++ R/print.R
@@ -21,7 +21,7 @@ print.gtsummary <- function(x, print_engine = NULL, ...) {
 
   # checking engine
   accepted_print_engines <-
-    c("gt", "kable", "kable_extra", "flextable", "tibble")
+    c("gt", "kable", "kable_extra", "flextable", "huxtable", "tibble")
   if (!rlang::is_string(print_engine) || !print_engine %in% accepted_print_engines) {
     stop(glue(
       "Select a valid print engine. ",
@@ -36,7 +36,7 @@ print.gtsummary <- function(x, print_engine = NULL, ...) {
     "kable" = as_kable(x),
     "flextable" = as_flex_table(x),
     "kable_extra" = as_kable_extra(x),
-    # "huxtable" = as_huxtable(x),
+    "huxtable" = as_hux_table(x),
     "tibble" = as_tibble(x)
   ) %>%
     print()
@@ -111,7 +111,7 @@ knit_print.gtsummary <- function(x, ...) {
     "kable" = as_kable(x),
     "flextable" = as_flex_table(x),
     "kable_extra" = as_kable_extra(x),
-    # "huxtable" = as_huxtable(x),
+    "huxtable" = as_hux_table(x),
     "tibble" = as_tibble(x)
   ) %>%
       knitr::knit_print()